MQTT/requirements-mqtt.txt
numpy==1.24.3
paho-mqtt
